Made </Header> a global component by moving it outside <Routes> in BrowserRouter and also added some UI to about page
